[CONFLICT] Remove dead code #6: Delete old recents implementation
Conflict with some recents components (Alternate)

Reverted “Ensuring that we reload a widget that was updated in another
orientation.”

Reverted “DO NOT MERGE. Don't use visibility flags to determine whether
to dismiss Recents. (Bug 18802531)

We can use our normal visibility check to dismiss Recents when the
screen is off,
since the system broadcast can occur after the activity is stopped.  We
should use
the same mechanism we use to test visibility when launching Recents and
just see
if it is the top most activity.”